fix: keep in-process executor within size limit
